🚚 Refactor TF coil geometry calculations by moving tf_global_geometry method to TFCoil class
