+++ process/sctfcoil.py
@@ -1143,8 +1143,6 @@ class SuperconductingTFCoil:
         """
         peaktfflag = 0
 
-        self.tf_global_geometry()
-
         # Calculation of the TF current from bt
         self.tf_current()
 
@@ -1481,78 +1479,6 @@ class SuperconductingTFCoil:
         ajwpro = ajcp * (acs / aturn)
 
         return ajwpro, vd
-
-    def tf_global_geometry(self):
-        """Subroutine for calculating the TF coil geometry
-        This includes:
-        - Overall geometry of coil (radii and toroidal planes area)
-        - Winding Pack NOT included
-        """
-
-        sctfcoil_module.theta_coil = np.pi / tfcoil_variables.n_tf_coils
-        sctfcoil_module.tan_theta_coil = np.tan(sctfcoil_module.theta_coil)
-
-        # TF coil inboard legs mid-plane cross-section area (WP + casing ) [m2]
-        if tfcoil_variables.i_tf_case_geom == 0:
-            # Circular front case
-            tfcoil_variables.tfareain = np.pi * (
-                build_variables.r_tf_inboard_out**2 - build_variables.r_tf_inboard_in**2
-            )
-        else:
-            # Straight front case
-            tfcoil_variables.tfareain = (
-                tfcoil_variables.n_tf_coils
-                * np.sin(sctfcoil_module.theta_coil)
-                * np.cos(sctfcoil_module.theta_coil)
-                * build_variables.r_tf_inboard_out**2
-                - np.pi * build_variables.r_tf_inboard_in**2
-            )
-
-        # Vertical distance from the midplane to the top of the tapered section [m]
-        if physics_variables.itart == 1:
-            sctfcoil_module.h_cp_top = (
-                build_variables.z_plasma_xpoint_upper + tfcoil_variables.dztop
-            )
-        # ---
-
-        # Outer leg geometry
-        # ---
-        # Mid-plane inner/out radial position of the TF coil outer leg [m]
-
-        sctfcoil_module.r_tf_outboard_in = (
-            build_variables.r_tf_outboard_mid - build_variables.dr_tf_outboard * 0.5e0
-        )
-        sctfcoil_module.r_tf_outboard_out = (
-            build_variables.r_tf_outboard_mid + build_variables.dr_tf_outboard * 0.5e0
-        )
-
-        # TF coil width in toroidal direction at inboard leg outer edge [m]
-        # ***
-        # Sliding joints geometry
-        if physics_variables.itart == 1 and tfcoil_variables.i_tf_sup != 1:
-            tfcoil_variables.tftort = (
-                2.0e0 * build_variables.r_cp_top * np.sin(sctfcoil_module.theta_coil)
-            )
-
-        # Default thickness, initially written for DEMO SC magnets
-        elif physics_variables.itart == 1 and tfcoil_variables.i_tf_sup == 1:
-            tfcoil_variables.tftort = (
-                2.0e0
-                * build_variables.r_tf_inboard_out
-                * np.sin(sctfcoil_module.theta_coil)
-            )
-        else:
-            tfcoil_variables.tftort = (
-                2.0e0
-                * build_variables.r_tf_inboard_out
-                * np.sin(sctfcoil_module.theta_coil)
-            )
-
-        # Area of rectangular cross-section TF outboard leg [m2]
-        tfcoil_variables.a_tf_leg_outboard = (
-            tfcoil_variables.tftort * build_variables.dr_tf_outboard
-        )
-        # ---
 
     def tf_current(self):
         """
